pyproject.toml
[project]
name = "fastapi-demoapp"
version = "0.1.0"
description = ""
authors = [
    {name = "mktakeda",email = "manav.khandurie@takeda.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.2,<46.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "strawberry-graphql[fastapi] (>=0.270.0,<0.271.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "opentelemetry-api (>=1.33.1,<2.0.0)",
    "opentelemetry-sdk (>=1.33.1,<2.0.0)",
    "opentelemetry-exporter-jaeger (>=1.21.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.54b1,<0.55)",
    "opentelemetry-instrumentation-asgi (>=0.54b1,<0.55)",
    "opentelemetry-exporter-otlp-proto-grpc (>=1.33.1,<2.0.0)",
    "opentelemetry-distro[otlp] (>=0.54b1,<0.55)",
    "opentelemetry-instrumentation (>=0.54b1,<0.55)",
    "opentelemetry-exporter-jaeger-thrift (>=1.21.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.33.1,<2.0.0)",
    "opentelemetry-instrumentation-logging (>=0.54b1,<0.55)",
    "loguru (>=0.7.3,<0.8.0)",
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
pytest = "^8.3.5"
isort = "^6.0.1"
pre-commit = "^4.2.0"
mypy = "^1.15.0"
pydantic = "^2.11.4"
bandit = "^1.8.3"
alembic = "^1.15.2"
pytest-cov = "^6.1.1"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
server = "src.main:run_server"


[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  tests/data/
  | profiling/
  | scripts/generate_schema.py  # Uses match syntax
)
'''

[tool.isort]
profile = "black"
line_length=88
multi_line_output = 3
include_trailing_comma = true
skip = [".venv","tests"]

